Added randomization functionality; lifted array state up to app
