+++ src/components/header-component/header.css
@@ -18,11 +18,12 @@ header > div > h2, button {
   border-radius: 0.75rem;
   padding: 0.25rem;
   cursor: pointer;
+  box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.08);
 }
 
 .clickable:hover {
   transform: scale(1.2);
-  box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.2);
+  box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.3);
   background-color: rgba(0, 0, 0, 0.05);
 }
